Move and exclude generated api files from linting. Make generate-api.js into a parameterized, reusable script.
