+++ package.json
@@ -12,14 +12,14 @@
     "buildBootstrap": "sass --style :compressed src/globalStyles/bootstrap-entry.scss src/globalStyles/prefixed-bootstrap.min.css",
     "heroku-postbuild": "npm run build && npm install http-server -g",
     "updateAPI": "npm run fetchAPI && npm run buildAPI",
-    "fetchAPI": "curl http://cbioportal-rc.herokuapp.com/api/v2/api-docs > src/shared/api/CBioPortalAPI-docs.json && curl http://cbioportal-rc.herokuapp.com/api/v2/api-docs?group=internal > src/shared/api/CBioPortalAPIInternal-docs.json",
-    "buildAPI": "node scripts/generate-api.js",
+    "fetchAPI": "curl http://cbioportal-rc.herokuapp.com/api/v2/api-docs > src/shared/api/generated/CBioPortalAPI-docs.json && curl http://cbioportal-rc.herokuapp.com/api/v2/api-docs?group=internal > src/shared/api/generated/CBioPortalAPIInternal-docs.json",
+    "buildAPI": "node scripts/generate-api.js src/shared/api/generated CBioPortalAPI CBioPortalAPIInternal",
     "updateHotspotsAPI": "npm run fetchHotspotsAPI && npm run buildHotspotsAPI",
-    "fetchHotspotsAPI": "curl http://cancerhotspots.org/v2/api-docs?group=cancer_hotspots > src/shared/api/CancerHotspotsAPI-docs.json",
-    "buildHotspotsAPI": "node scripts/generate-hotspots-api.js",
+    "fetchHotspotsAPI": "curl http://cancerhotspots.org/v2/api-docs?group=cancer_hotspots > src/shared/api/generated/CancerHotspotsAPI-docs.json",
+    "buildHotspotsAPI": "node scripts/generate-api.js src/shared/api/generated CancerHotspotsAPI",
     "test": "karma start karma.conf.js",
     "test:watch": "npm run test -- --watch",
-    "lint": "find . -name '*.tsx' | xargs tslint",
+    "lint": "tslint src/**/*.ts src/**/*.tsx --exclude src/shared/api/generated/**",
     "eslint": "eslint src/** --ext .js --ext .jsx --cache",
     "syncmock": "node src/test/fetchMockData.js --diff",
     "validate": "npm ls"
