+++ scripts/generate-api.js
@@ -1,16 +1,12 @@
-// output to src/shared/api
-var fs = require('fs');
-var path = require('path');
-var request = require('request');
-var CodeGen = require('swagger-js-codegen').CodeGen;
+const fs = require('fs');
+const path = require('path');
+const request = require('request');
+const CodeGen = require('swagger-js-codegen').CodeGen;
 
-var swagger = JSON.parse(fs.readFileSync('src/shared/api/CBioPortalAPI-docs.json'));
-var tsSourceCode = CodeGen.getTypescriptCode({className: 'CBioPortalAPI', swagger});
-fs.writeFileSync('src/shared/api/CBioPortalAPI.ts', tsSourceCode);
-
-
-swagger = JSON.parse(fs.readFileSync('src/shared/api/CBioPortalAPIInternal-docs.json'));
-tsSourceCode = CodeGen.getTypescriptCode({className: 'CBioPortalAPIInternal', swagger});
-fs.writeFileSync('src/shared/api/CBioPortalAPIInternal.ts', tsSourceCode);
-
-
+const [node, script, folder, ...classNames] = process.argv;
+for (const className of classNames)
+{
+	const swagger = JSON.parse(fs.readFileSync(path.join(folder, `${className}-docs.json`)));
+	const tsSourceCode = CodeGen.getTypescriptCode({className, swagger});
+	fs.writeFileSync(path.join(folder, `${className}.ts`), tsSourceCode);
+}
