+++ src/pages/patientView/clinicalInformation/ClinicalInformationPatientTable.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import {ClinicalData} from "../../../shared/api/CBioPortalAPI";
+import {ClinicalData} from "../../../shared/api/generated/CBioPortalAPI";
 import {Table as DataTable} from "reactableMSK";
 import TableHeaderControls from "shared/components/tableHeaderControls/TableHeaderControls";
