+++ src/pages/patientView/clinicalInformation/getClinicalInformationData.ts
@@ -1,8 +1,8 @@
 import * as queryString from "query-string";
 import * as _ from 'lodash';
-import CBioPortalAPI from "../../../shared/api/CBioPortalAPI";
+import CBioPortalAPI from "../../../shared/api/generated/CBioPortalAPI";
 import { ClinicalDataBySampleId } from "../../../shared/api/api-types-extended";
-import {ClinicalData} from "../../../shared/api/CBioPortalAPI";
+import {ClinicalData} from "../../../shared/api/generated/CBioPortalAPI";
 import {ClinicalInformationData} from "../Connector";
 //import { getTreeNodesFromClinicalData, PDXNode } from './PDXTree';
 //import sampleQuery from 'shared/api/mock/Samples_query_patient_P04.json';
