+++ src/pages/patientView/clinicalInformation/lib/convertSamplesData.spec.ts
@@ -1,5 +1,5 @@
 import mockClinicalData from "../../../../shared/api/mock/Clinical_data_study_ucec_tcga_pub.json";
-import { ClinicalData } from "../../../../shared/api/CBioPortalAPI";
+import { ClinicalData } from "../../../../shared/api/generated/CBioPortalAPI";
 import { default as convertSampleData, IConvertedSamplesData } from "./convertSamplesData";
 import { groupByEntityId } from '../getClinicalInformationData';
 import { ClinicalDataBySampleId } from "../../../../shared/api/api-types-extended";
