+++ src/pages/patientView/genomicOverview/Tracks.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import * as _ from 'lodash';
 import * as tracksHelper from './tracksHelper'
-import {CopyNumberSeg, Mutation} from 'shared/api/CBioPortalAPI';
+import {CopyNumberSeg, Mutation} from 'shared/api/generated/CBioPortalAPI';
 import SampleManager from "../sampleManager";
 import {ClinicalDataBySampleId} from "../../../shared/api/api-types-extended";
