+++ src/pages/patientView/genomicOverview/tracksHelper.ts
@@ -3,7 +3,7 @@ import $ from 'jquery';
 import * as _ from 'lodash';
 import 'qtip2';
 import 'qtip2/dist/jquery.qtip.css';
-import { Mutation } from "shared/api/CBioPortalAPI";
+import { Mutation } from "shared/api/generated/CBioPortalAPI";
 
 export function GenomicOverviewConfig(nRows: any,width: any) {
     let sel: any = {};
