+++ src/pages/patientView/mutation/column/AlleleFreqColumnFormatter.tsx
@@ -6,7 +6,7 @@ import DefaultTooltip from "shared/components/DefaultTooltip";
 import {compareNumberLists} from '../../../../shared/lib/SortUtils';
 import 'rc-tooltip/assets/bootstrap_white.css';
 import {MutationTableRowData} from "../../../../shared/components/mutationTable/IMutationTableProps";
-import {Mutation} from "../../../../shared/api/CBioPortalAPI";
+import {Mutation} from "../../../../shared/api/generated/CBioPortalAPI";
 
 export default class AlleleFreqColumnFormatter {
     static barWidth = 6;
