--- src/shared/api/CBioPortalAPI-docs.internal.json
@@ -1,1 +1,0 @@
-{"swagger":"2.0","info":{"description":"A web service for supplying JSON formatted data to cBioPortal clients.","version":"1.0 (beta)","title":"cBioPortal web API","termsOfService":"www.cbioportal.org","contact":{"name":"cBioPortal","url":"www.cbioportal.org","email":"cbioportal@googlegroups.com"},"license":{"name":"License","url":"https://github.com/cBioPortal/cbioportal/blob/master/LICENSE"}},"host":"cbioportal-rc.herokuapp.com","basePath":"/api","tags":[{"name":"Variant Counts","description":" "},{"name":"mRNA Percentile","description":" "},{"name":"Significantly Mutated Genes","description":" "},{"name":"Cosmic Counts","description":" "}],"paths":{"/cosmic-counts/fetch":{"post":{"tags":["Cosmic Counts"],"summary":"Get counts within the COSMIC database by keywords","operationId":"fetchCosmicCountsUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"keywords","description":"List of keywords","required":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/CosmicMutation"}}}}}},"/genetic-profiles/{geneticProfileId}/mrna-percentile/fetch":{"post":{"tags":["mRNA Percentile"],"summary":"Get mRNA expression percentiles for list of genes for a sample","operationId":"fetchMrnaPercentileUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"geneticProfileId","in":"path","description":"Genetic Profile ID e.g. acc_tcga_rna_seq_v2_mrna","required":true,"type":"string"},{"name":"sampleId","in":"query","description":"Sample ID e.g. TCGA-OR-A5J2-01","required":true,"type":"string"},{"in":"body","name":"entrezGeneIds","description":"List of Entrez Gene IDs","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int32"}}}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/MrnaPercentile"}}}}}},"/genetic-profiles/{geneticProfileId}/variant-counts/fetch":{"post":{"tags":["Variant Counts"],"summary":"Get counts of specific variants within a mutation genetic profile","operationId":"getVariantCountsUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"geneticProfileId","in":"path","description":"Genetic Profile ID e.g. acc_tcga_mutations","required":true,"type":"string"},{"in":"body","name":"variantCountIdentifiers","description":"List of Entrez Gene IDs","required":true,"schema":{"type":"array","items":{"$ref":"#/definitions/VariantCountIdentifier"}}}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/VariantCount"}}}}}},"/studies/{studyId}/significantly-mutated-genes":{"get":{"tags":["Significantly Mutated Genes"],"summary":"Get significantly mutated genes in a study","operationId":"getSignificantlyMutatedGenesUsingGET","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"studyId","in":"path","description":"Study ID e.g. acc_tcga","required":true,"type":"string"},{"name":"projection","in":"query","description":"Level of detail of the response","required":false,"type":"string","default":"SUMMARY","enum":["ID","SUMMARY","DETAILED","META"]},{"name":"pageSize","in":"query","description":"Page size of the result list","required":false,"type":"integer","default":1000,"format":"int32"},{"name":"pageNumber","in":"query","description":"Page number of the result list","required":false,"type":"integer","default":0,"format":"int32"},{"name":"sortBy","in":"query","description":"Name of the property that the result list is sorted by","required":false,"type":"string","enum":["entrezGeneId","hugoGeneSymbol","rank","numberOfMutations","pValue","qValue"]},{"name":"direction","in":"query","description":"Direction of the sort","required":false,"type":"string","default":"ASC","enum":["ASC","DESC"]}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/MutSig"}}}}}}},"definitions":{"VariantCountIdentifier":{"type":"object","properties":{"entrezGeneId":{"type":"integer","format":"int32"},"keyword":{"type":"string"}}},"CosmicMutation":{"type":"object","properties":{"cosmicMutationId":{"type":"string"},"count":{"type":"integer","format":"int32"},"keyword":{"type":"string"},"proteinChange":{"type":"string"}}},"VariantCount":{"type":"object","properties":{"entrezGeneId":{"type":"integer","format":"int32"},"geneticProfileId":{"type":"string"},"keyword":{"type":"string"},"numberOfSamples":{"type":"integer","format":"int32"},"numberOfSamplesWithKeyword":{"type":"integer","format":"int32"},"numberOfSamplesWithMutationInGene":{"type":"integer","format":"int32"}}},"MrnaPercentile":{"type":"object","properties":{"entrezGeneId":{"type":"integer","format":"int32"},"geneticProfileId":{"type":"string"},"percentile":{"type":"number","format":"double"},"sampleId":{"type":"string"},"zScore":{"type":"number","format":"double"}}},"MutSig":{"type":"object","properties":{"entrezGeneId":{"type":"integer","format":"int32"},"hugoGeneSymbol":{"type":"string"},"numberOfMutations":{"type":"integer","format":"int32"},"pValue":{"type":"number","format":"double"},"qValue":{"type":"number","format":"double"},"rank":{"type":"integer","format":"int32"},"studyId":{"type":"string"}}}}}
