+++ src/shared/api/api-types-extended.ts
@@ -1,4 +1,4 @@
-import { ClinicalData } from "./CBioPortalAPI";
+import { ClinicalData } from "./generated/CBioPortalAPI";
 
 export type ClinicalDataBySampleId = {
     id: string;
