+++ src/shared/api/cbioportalClientInstance.ts
@@ -1,3 +1,3 @@
-import CBioPortalAPI from "./CBioPortalAPI";
+import CBioPortalAPI from "./generated/CBioPortalAPI";
 const client = new CBioPortalAPI(`//${(window as any)['__API_ROOT__']}`);
 export default client;
