+++ src/shared/api/cbioportalInternalClientInstance.ts
@@ -1,3 +1,3 @@
-import CBioPortalAPIInternal from "./CBioPortalAPIInternal";
-const client = new CBioPortalAPIInternal(`//${(window as any)['__API_ROOT__']}`);
-export default client;
+import CBioPortalAPIInternal from "./generated/CBioPortalAPIInternal";
+const internalClient = new CBioPortalAPIInternal(`//${(window as any)['__API_ROOT__']}`);
+export default internalClient;
