+++ src/shared/components/mutationTable/column/GeneColumnFormatter.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {Td} from 'reactable';
 import {IColumnFormatterData} from "shared/components/enhancedReactTable/IColumnFormatter";
 import {MutationTableRowData} from "../IMutationTableProps";
-import {Mutation, Gene} from "shared/api/CBioPortalAPI";
+import {Mutation, Gene} from "shared/api/generated/CBioPortalAPI";
 
 /**
  * @author Selcuk Onur Sumer
