+++ src/shared/components/mutationTable/column/ProteinChangeColumnFormatter.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import {Td} from 'reactable';
 import {IColumnFormatterData} from "shared/components/enhancedReactTable/IColumnFormatter";
 import {MutationTableRowData} from "../IMutationTableProps";
-import {Mutation} from "shared/api/CBioPortalAPI";
+import {Mutation} from "shared/api/generated/CBioPortalAPI";
 
 /**
  * @author Selcuk Onur Sumer
